requirements.txt
Django~=2.2.4
django-crispy-forms
python-slugify
pyqrcode
pyzbar
pyzbar[scripts]
pypng
image
